requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argcomplete==3.6.3
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
blivet==3.12.1
Brlapi==0.8.7
cachetools==6.2.4
certifi==2025.7.9
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.7
cockpit @ file:///builddir/build/BUILD/cockpit-354-build/cockpit-354/tmp/wheel/cockpit-354-py3-none-any.whl#sha256=1d4fb1517c54d5d11d999be3583ea5fbf3d3d011cb789fb717daae0672c38816
comm==0.2.3
contourpy==1.3.3
crypt_r==3.13.1
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-python==1.4.0
dbus_next==0.2.3
debugpy==1.8.19
decorator==5.2.1
distro==1.9.0
dnf==4.24.0
dnspython==2.8.0
email-validator==2.3.0
evdev==1.9.1
executing==2.2.1
fastapi==0.128.0
fedora-third-party==0.10
file-magic==0.4.0
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
idna==3.10
imbalanced-learn==0.14.1
input-remapper==2.2.0
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
iso639==0.1.4
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
langtable==0.0.69
libcomps==0.1.23
libdnf==0.75.0
libdnf5==5.2.17.0
louis==3.33.0
lxml==6.0.1
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
narwhals==2.15.0
nest-asyncio==1.6.0
nftables==0.1
numpy==2.4.1
olefile==0.47
packaging==25.0
pandas==2.3.3
parso==0.8.5
perf==0.1
pexpect==4.9.0
pid==3.0.4
pillow==11.3.0
platformdirs==4.5.1
ply==3.11
productmd==1.49
progressbar2==4.5.0
prompt_toolkit==3.0.52
protobuf==6.33.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
pyarrow==22.0.0
PyAudio==0.2.13
pycairo==1.28.0
pycparser==2.22
pycups==2.0.4
pydantic==2.12.5
pydantic_core==2.41.5
pydbus==0.6.0
pydeck==0.9.1
pyenchant==3.2.2
Pygments==2.19.2
PyGObject==3.54.5
pyinotify==0.9.6
pykickstart @ file:///builddir/build/BUILD/pykickstart-3.66-build/pykickstart-3.66
pyparsing==3.3.1
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.2.0
python-dateutil==2.9.0.post0
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
python-utils==3.9.1
pytz==2025.2
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.18.5-build/kernel-6.18.5/linux-6.18.5-200.fc43.x86_64/tools/net/ynl
pyzmq==27.1.0
RapidFuzz==3.12.2
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
requests-file==2.0.0
requests-ftp==0.3.1
rpds-py==0.27.0
rpm==6.0.1
rpmautospec==0.8.3
rpmautospec-core==0.1.5
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
selinux @ file:///builddir/build/BUILD/libselinux-3.9-build/libselinux-3.9/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.9-build/selinux-3.9/python/sepolicy
setools==4.6.0
setuptools==78.1.1
simpleline==1.9.0
six==1.17.0
sklearn-compat==0.1.5
smmap==5.0.2
sniffio==1.3.1
sos==4.10.2
soupsieve==2.8
stack-data==0.6.3
starlette==0.50.0
streamlit==1.53.0
systemd-python==235
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
wcwidth==0.2.14
xkbregistry==0.3
